Stage number is displayed
